refactor: P2 server split 2/N — extract /api/channels + /api/bridge route domain
Pure cut-paste, zero behavior change:

- server/routes/channels.ts: all 10 blocks covering /api/channels* (list,
  default-target, test, health, feishu QR register/status, wechat iLink QR
  login/status/status, runs log) and /api/bridge/messages become
  handleChannelsRoutes(...): Promise<boolean>
- Mutable module state that QR flows reassign (config, wechatChannel) is
  injected via getter/setter pairs on ChannelsRouteContext instead of being
  captured by value; dispatcher/bridgeToken get getters for the same reason
  (they are initialized lazily during bootstrap)
- server.ts keeps one delegation call at the first block's original
  position; drops the now-unused feishu-registration and bridge-server
  imports
- Smoke test gains two channels assertions (empty channel list 200, bridge
  404 when unconfigured)

server.ts 5155 -> 4959 lines.
